// Neptune.Web/src/app/shared/generated/model/treatment-bmp-create-dto.ts
export class TreatmentBMPCreateDto { 
    TreatmentBMPName: string;
    TreatmentBMPTypeID: number;
    StormwaterJurisdictionID: number;
    OwnerOrganizationID?: number | null;
    YearBuilt?: number | null;
    SystemOfRecordID?: string | null;
    WaterQualityManagementPlanID: number;
    TreatmentBMPLifespanTypeID: number;
    TreatmentBMPLifespanEndDate?: string | null;
    SizingBasisTypeID: number;
    TrashCaptureStatusTypeID: number;
    TrashCaptureEffectiveness?: number | null;
    RequiredFieldVisitsPerYear?: number | null;
    RequiredPostStormFieldVisitsPerYear?: number | null;
    Notes?: string | null;
    Latitude: number;
    Longitude: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPCreateDtoForm { 
    TreatmentBMPName: FormControl<string>;
    TreatmentBMPTypeID: FormControl<number>;
    StormwaterJurisdictionID: FormControl<number>;
    OwnerOrganizationID?: FormControl<number>;
    YearBuilt?: FormControl<number>;
    SystemOfRecordID?: FormControl<string>;
    WaterQualityManagementPlanID: FormControl<number>;
    TreatmentBMPLifespanTypeID: FormControl<number>;
    TreatmentBMPLifespanEndDate?: FormControl<string>;
    SizingBasisTypeID: FormControl<number>;
    TrashCaptureStatusTypeID: FormControl<number>;
    TrashCaptureEffectiveness?: FormControl<number>;
    RequiredFieldVisitsPerYear?: FormControl<number>;
    RequiredPostStormFieldVisitsPerYear?: FormControl<number>;
    Notes?: FormControl<string>;
    Latitude: FormControl<number>;
    Longitude: FormControl<number>;
}